First commit; combining sandbox projects, compiles
Multiple different projects for playing with various convolution,
morphology, and workbench techniques using CUDA.  Now I've decided
to reorganize everything into a more formal project that includes
everything in with better coding standards and organization.

This commit doesn't include everything.  In fact, it only includes
skeletons for a lot of things.  I need to make sure I can get the
CUDA kernel code compiling and working before I start adding anything
that uses it.  CUDA is a royal pain in the ass to debug, and I've
gotten it working before, so I need to get the most basic stuff
working in the new organization scheme, and pull in everything else
incrementally, verifying that I didn't break anything at any step.
CUDA is ridiculously stubborn, and refuses to tell you where your
errors are, so this is the best way to do it, as far as I can tell.