CUDA works, test passed!  17x17 dilate & erode
View the png file.  I used a very unique 17x17 structuring element to
dilate and erode the salt image.  This image is constrcuted so it has
white salt on black background, and black salt on white background.
This helps me confirm that dilate and erode are doing what they are
supposed to.  Notice that the shape is mirrored in the resulting images,
that is expected.  Also notice that the edges are pulled inwards on the
eroded image, that is also expected but may not be desirable.  Perhaps
the edge behavior should be customizable.