Functional and optimizaed 3x3 structuring elements
Since 3x3 SEs are so small, we created a macro which actually rewrites
the Morph_Generic_Kernel function, but removing all the loops and
hardcoding the structuring elements inside the kernel code.  Haven't
done any timings yet, but this is likely to give a significant increase
in speed, though it might use more registers

dispGpuSoln.m is a MATLAB script which reads all the output from the
unit test and displays the I/O.  As of this commit, we can verify that
morphology with arbitrary structuring elements along with 3x3 optimized
are all working.