Host&Device memory management w/ cudaImage classes
Added cudaImageHost and cudaImageDevice which automate the allocation
and freeing of host and device memory.  The user only needs to
specify the 2D dimensions of the desired chunk of memory.  The DEVICE
memory is tracked by a static list of pointers to every DEVICE
allocation made, so that the memory utilization can be tracked and
queried at any time.  I didn't feel it was necessary to track the HOST
memory...

Anyone who is getting into CUDA for the first time might benefit from
these classes, since memory management can be archaic and the source of
many headaches.

Also added some GPU timing code using CUDA events.  These GPU times only
measure GPU clock cycles, and so will necessarily return a value less
than wall-time.  This is useful for benchmarking the GFLOPS of your GPU
and the bandwidth of the PCI-e connection.  To measure actual wall-time,
the CPU timer (cpuStartTime/cpuStopTime in cudaConvUtilities) should do
a pretty good job.

With all of the memory management code written, and working morphology
of arbitrary and 3x3 structuring elements, next I will complete the
workbench and start using it for things.  My first task will be to solve
a maze based on a clean picture of it...