IWB buffer access improved and some methods tested
I revised the way ImageWorkbench buffers are accessed.  I originally
used a "BUFFER TYPE" and a buffer index, but it's bulky, ugly, and
really a pain to maintain.  Now I just use a single index where A and B
are #define'd values, and anything 1+ accesses a (1-indexed) extra
buffer, and IWB can use -1 or lower to access the temporary buffers.

Also got a bunch of IWB functions working.  In general, buffer flipping
works, and basic morphological operations work.  However, it seems that
my subtract kernel doesn't, which prevents me from doing some of the
cooler things possible with morphology.  Working on it...