Fixed MASSIVE speed bug!  See TimingTest.txt
I should've guessed I had done something stupid like this, when the code
was about 3x-4x slower than the float version!  I was allocating 8192
bytes of shared memory with
               __shared__ int shmem[8192];
instead of
               __shared__ char shmem[8192];

So I was reserving 4x the amount of shared memory I actually needed,
which reduced my occupancy to 1/6!  That's what I get for not
profiling it earlier...

Updated timings for the GTX 460 are:

	Morph  512x 512 image with optimized 3x3:     0.22 ms -- 4589.3 FPS)
	Morph  512x 512 image with  3x 3 struct elt:  0.34 ms -- 2899.4 FPS)
	Morph  512x 512 image with  5x 5 struct elt:  0.47 ms -- 2118.6 FPS)
	Morph  512x 512 image with  7x 7 struct elt:  0.71 ms -- 1415.0 FPS)
	Morph  512x 512 image with  9x 9 struct elt:  0.95 ms -- 1053.7 FPS)

Also changed the test maze to something that can be easily viewed
without zooming.  It's a much simpler maze, but it's also easier and
faster to execute the unit tests on (now completed in less than 1s on
the GTX 460).

Next task will be to convert everything back to row-major so that it is
aligned with the rest of the world (except CUDA itself), and then find
the SEG FAULT which keeps crashing the program during destruction.

Following that, I can start expanding more.  Perhaps gray-scale
morphology next...