Everything in row-major format, README updated
CUDA blocks and grids are always accessed with col-major (c,r) indexing
but we prefer row-major (r,c) indexing.  I originally conceded to using
col-major to be consistent with CUDA, but became inconsistent with
everything else in my life.  Therefore, it's changed to
   BLOCK.x ~ GRID.x ~ Image Row
   BLOCK.y ~ GRID.y ~ Image Column

Also updated the README and attached the full output of the unit tests
for a GTX 460 card.