Geforce 210 sucks!...or I suck programming them...
I was able to compile and run the code on my extremely-low-end Geforce
210 graphics card.  Seriously, I got it for $20, and it's the lowest
200-series card out there.  The results were less than flattering...

To see the complete output for both video cards, vimdiff the two files:
   Output_GTX460.txt      and
   Output_Geforce210.txt

The code worked and passed the unit tests, but generally was 1/5 to 1/30
the speed of the GTX 460.  HOWEVER, this may not be the card's fault
(entirely)... I wrote the algorithms to be optimized for the Fermi
architecture, with no regard for what other architectures might suffer.
Also, even at 1/30 the speed, it's still faster than my CPU by a factor
of 5x.

In the future, the following preprocessor construct can be used to
branch the code.  The appendices in the the CUDA 3.1 programming guide
should give enough information to know how to adapt various code for the
different architectures.  For reference, the GTX 460 is compute
capability 2.1, and the Geforce 210 is capability 1.2.

    // Device code path for compute capability 1.0+
    // Device code path for compute capability 2.0+
    // Host code path